requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-bcrypt==1.0.1
flask-cors==4.0.0
psycopg2-binary==2.9.9
redis==5.0.1
gunicorn==21.2.0
werkzeug==3.0.1
PyMuPDF==1.23.5
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
